+++ Minecraft/Networking/Packets/Play/Serverbound/KeepAliveServerboundPacket.swift
@@ -10,9 +10,9 @@ import Foundation
 struct KeepAliveServerBoundPacket: ServerboundPacket {
   static let id: Int = 0x10
   
-  var keepAliveId: Int64
+  var keepAliveId: Int
   
   func writePayload(to writer: inout PacketWriter) {
-    writer.writeLong(keepAliveId)
+    writer.writeLong(Int64(keepAliveId))
   }
 }
